requirements.txt
pytest
playwright
pytest-playwright
requests
pytest-html
pyyaml
mysql-connector-python  # For MySQL, or change based on DB